Add feerate to `LatestCounterpartyCommitmentTXInfo`
Should we expose full `CommitmentTransaction` objects from monitor
updates we will need to accurately report the feerate.

The feerate is not totally necessary (depending on what other values you
possess) to have in order to reform the counterparty commitment
transaction from a monitor update, so in the case where we only expose
revokeable output data, this could be removed.