Add local and remote values on `LatestCounterpartyCommitmentTXInfo`
These are the minimum fields that must be added to the monitor update in
order to allow the `ChannelMonitor` to reconstruct the counterparty's
commitment transaction, and find the value and output index of the
revokeable output we'll claim in the justice transaction.

With only the HTLC outputs currently available in the monitor update, we
can tell how much of the channel balance is in-flight and towards which
side, however it doesn't tell us the amount that resides on either side.

Because of dust, we can't reliably derive the remote value from the
local value and visa versa.