Approach 1: Get counterparty commitment txs from monitor update
This is one approach for allowing a user to build watchtower
functionality into the persistence pipeline.

Pros: Getting commitment transactions from the monitor update is a relatively
intuitive interface, and getting a presigned justice tx for that
commitment is also pretty intuitive.

Cons: However, for a watchtower, in between seeing a counterparty commitment
tx and building the justice tx, you'd have to store the whole commitment
tx, which is a lot of unnecessary data. To find the revokeable output in
the commitment, we need the counterparty's commitment point for that tx,
which is available in
`ChannelMonitorUpdateStep::LatestCounterpartyCommitmentTXInfo`. However,
because it isn't available to the user, the `ChannelMonitor` needs to
take the whole update to find any revokeable outputs (which ends up
looking more similar to approach 2 in the next commit).

Building and signing a justice tx from a commitment in one go also
doesn't allow for that much flexibility if a user wanted to make
modifications to the justice tx.