Approach 2: Get revoked output data from update, build/sign separately
This is a second approach for building watchtower functionality.

This exposes the minimal information needed to claim a revoked
output. We also wouldn't need to add the feerate on the monitor update
if we only exposed this.

This also allows for more flexibility in justice tx construction, as we
only provide a helper to build the tx, but signing is independent.