Allow providing an unsigned funding (splice) transaction
When providing the funding transaction for the new channel, we're
providing our closing transaction for our old channel. Providing this
transaction is only relevant to getting our new channel's intitial
commitment signed. We handle signing this transaction in our
`closing_signed` logic. So here we provide an extra method on
`ChannelManager` to let a splicing user skip signing.